Fix rank in models.

+++ db/migrate/20090713152149_create_episodes.rb
@@ -14,7 +14,7 @@ class CreateEpisodes < ActiveRecord::Migration
       t.integer :losing_votes_count,  :default => 0
       t.integer :total_votes_count,   :default => 0
       t.float   :total_votes_share,   :default => 0.0
-      t.integer :rank,                :default => 0
+      t.float   :rank,                :default => 0.0
       t.timestamps
     end
     add_index :episodes, :director_id
