+++ db/migrate/20090713152154_create_seasons.rb
@@ -9,7 +9,7 @@ class CreateSeasons < ActiveRecord::Migration
       t.integer :losing_votes_count,  :default => 0
       t.integer :total_votes_count,   :default => 0
       t.float   :total_votes_share,   :default => 0.0
-      t.integer :rank,                :default => 0
+      t.float   :rank,                :default => 0.0
       t.timestamps
     end
     add_index :seasons, :total_votes_share
